+++ frontend/src/services/applyForm/validation.ts
@@ -1,0 +1,5 @@
+import { FormSchema } from "./types";
+
+export const validateJsonSchema = (schema: FormSchema) => {
+    return schema;
+}
